fix: externalRepresentation condition to validate if key is private should be d not prime
I got to this issue recently where the library was giving me a public key DER when the key had a d and no primes,
this is misleading since an RSA key can have d and no primes.
This function should acess if the key is private through the d, if it requires the primes it should return an error.

As part of this PR I also made the primes public, they are already a let, making them public makes the API more
accessible to deal for example if I want to build a JWK with primes included.